+++ foundry/src/MarketMap.sol
@@ -6,89 +6,58 @@ import "./MarketInterface.sol";
 
 /// @title MarketMap
 /// @author mozrt (mozrt.eth)
-/// @notice This contract stores and manages simple ERC721 (NFT) buy and sell orders for Seaport onchain.
-
-struct UniqueOrderParameters {
-    uint32 amount;
-    uint32 startTime;
-    uint32 endTime; 
-    uint16 salt; 
-    bytes1 vSignature;
-    bytes32 rSignature;
-    bytes32 sSignature;
-}
+/// @notice This contract stores and manages simple Seaport buy and sell orders onchain.
 
 contract MarketMap {
 
     address immutable seaportAddress;
     ISeaport immutable seaport;
-    bytes32 immutable conduitKey;
+    address immutable openseaAddress;
+    uint256 immutable openseaFee;
+    bytes32 immutable openseaConduitKey;
 
     /// @notice Creates a new MarketMap contract.
     /// @param _seaportAddress The address of the Seaport contract.
-    /// @param _conduitKey The key used for the conduit.
+    /// @param _openseaAddress The address of the Opensea contract.
+    /// @param _openseaFee The fee charged by Opensea for a sale of 1 ETH.
+    /// @param _openseaConduitKey The key for the Opensea Conduit.
     constructor (
         address _seaportAddress, 
-        bytes32 _conduitKey
+        address _openseaAddress,
+        uint256 _openseaFee,
+        bytes32 _openseaConduitKey
         ) {
         seaportAddress = _seaportAddress;
         seaport = ISeaport(_seaportAddress);
-        conduitKey = _conduitKey;
+        openseaAddress = _openseaAddress;
+        openseaFee = _openseaFee;
+        openseaConduitKey = _openseaConduitKey;
     }
 
+    /// @notice Mapping to store unique order parameters for each token ID of a given address.
     mapping(address => mapping(uint256 => UniqueOrderParameters)) public orders;
+
+    /// @notice Mapping to store additional recipients (e.g. royalty recipients) for each token ID of a given address.
+    mapping(address => mapping(uint256 => mapping(uint8 => AdditionalRecipient))) public additionalRecipients;
+
+    /// @notice Mapping to store a list of all token IDs with orders for each NFT collection.
     mapping(address => uint256[]) private tokenIdsWithOrders;
+
+    /// @notice Mapping to store the index of each token ID in tokenIdsWithOrders.
     mapping(address => mapping(uint256 => uint256)) private tokenIdToIndex;
 
     /// @notice Creates a sell order after ensuring that it is in the valid format and the token is approved for Seaport.
     /// @param order The order parameters in Seaport's format.
-    function sell(
-        Order memory order 
-    ) public {
-        uint256 amount = order.parameters.consideration[0].startAmount;
-        require(amount%1e13 == 0 && amount < 1e22, "Wei amount cannot have any digits other than 0 for its first 13 digits and cannot exceed 10,000 ETH");
+    /// @param openseaSignature Optional signature from Opensea, if provided it will submit an order to Opensea.
+    function sell(Order memory order, bytes memory openseaSignature) public {
+        validateOrderParameters(order);
 
-        uint256 endTime = order.parameters.endTime;
-        require(endTime < 4.2*1e9, "End time must be smaller than 4.2 billion");
+        Order[] memory orderList = processOpenseaSignature(order, openseaSignature);
 
-        uint256 salt = order.parameters.salt;
-        require(salt < 65000, "Salt must be less than 65,000");
-
-        address approvedAddress = IERC721(order.parameters.offer[0].token).getApproved(order.parameters.offer[0].identifierOrCriteria);
-        require(approvedAddress == seaportAddress, "Seaport not approved to transfer token");
-
-        Order[] memory orderList = new Order[](1);
-        orderList[0] = order;
         bool validated = seaport.validate(orderList);
-        require(validated == true, "Invalid Order");
+        require(validated == true, "Invalid order");
 
-        bytes1 v;
-        bytes32 r;
-        bytes32 s;
-
-        bytes memory signature = order.signature;
-
-        assembly {
-            v := mload(add(signature, 0x20))
-            r := mload(add(signature, 0x21))
-            s := mload(add(signature, 0x41))
-        }
-
-        UniqueOrderParameters memory uniqueOrderParameters = UniqueOrderParameters({
-            amount: uint32(order.parameters.consideration[0].startAmount/1e13),
-            startTime: uint32(order.parameters.startTime),
-            endTime: uint32(order.parameters.endTime),
-            salt: uint16(order.parameters.salt),
-            vSignature: v,
-            rSignature: r,
-            sSignature: s
-        });
-
-        address token = order.parameters.offer[0].token;
-        uint256 tokenId = order.parameters.offer[0].identifierOrCriteria;
-
-        orders[token][tokenId] = uniqueOrderParameters;
-        addTokenId(token, tokenId);
+        storeOrderAndRecipients(order);
     }
  
     /// @notice Triggers the purchase of an NFT listed for sale on MarketMap.
@@ -103,7 +72,7 @@ contract MarketMap {
         AdvancedOrder memory advancedOrder = composeOrder(token, tokenId);
         CriteriaResolver[] memory emptyCriteriaResolver = new CriteriaResolver[](0);
 
-        bool fulfilled = seaport.fulfillAdvancedOrder{value: msg.value}(advancedOrder, emptyCriteriaResolver, conduitKey, recipient);
+        bool fulfilled = seaport.fulfillAdvancedOrder{value: msg.value}(advancedOrder, emptyCriteriaResolver, bytes32(0), recipient);
 
         if (fulfilled) {
             delete orders[token][tokenId]; 
@@ -152,7 +121,201 @@ contract MarketMap {
         require(cancelled == true, "Order couldn't be cancelled");
 
         delete orders[token][tokenId];
+
+        uint256 considerations;
+        uint8 i = 1;	
+        while (additionalRecipients[token][tokenId][i].amount > 0) {
+            considerations++;
+            i++;
+            delete additionalRecipients[token][tokenId][i];
+        }
+        
         removeTokenId(token, tokenId);
+    }
+
+    /// @notice Retrieves the status of a given order.
+    /// @param token The address of the token involved in the order.
+    /// @param tokenId The token ID involved in the order.
+    /// @return isActive A boolean value indicating whether the order is active or not.
+    function getOrderStatus(address token, uint256 tokenId) public view returns (bool isActive) {
+        AdvancedOrder memory advancedOrder = composeOrder(token, tokenId);
+        OrderComponents memory orderComponents = OrderComponents({
+            offerer: advancedOrder.parameters.offerer,
+            zone: advancedOrder.parameters.zone,
+            offer: advancedOrder.parameters.offer,
+            consideration: advancedOrder.parameters.consideration,
+            orderType: advancedOrder.parameters.orderType,
+            startTime: advancedOrder.parameters.startTime,
+            endTime: advancedOrder.parameters.endTime,
+            zoneHash: advancedOrder.parameters.zoneHash,
+            salt: advancedOrder.parameters.salt,
+            conduitKey: advancedOrder.parameters.conduitKey,
+            counter: seaport.getCounter(advancedOrder.parameters.offerer)
+        });
+        bytes32 orderHash = seaport.getOrderHash(orderComponents);
+        (isActive,,,) = seaport.getOrderStatus(orderHash);
+        return isActive;
+    }
+
+    /// @notice Pre-validates the order parameters to ensure they can be stored without issues.
+    /// @param order The order parameters in Seaport's format.
+    function validateOrderParameters(Order memory order) internal view {
+        uint256 amount = order.parameters.consideration[0].startAmount;
+        require(amount % 1e3 == 0 && amount < 1.84 * 1e22, "Wei amount cannot have any digits other than 0 for its first 3 digits and cannot exceed 18,400 ETH");
+
+        uint256 endTime = order.parameters.endTime;
+        require(endTime < 4.2 * 1e9, "End time must be smaller than 4.2 billion");
+
+        uint256 salt = order.parameters.salt;
+        require(salt < 65000, "Salt must be less than 65,000");
+
+        address approvedAddress = IERC721(order.parameters.offer[0].token).getApproved(order.parameters.offer[0].identifierOrCriteria);
+        require(approvedAddress == seaportAddress, "Seaport not approved to transfer token");
+    }
+
+    /// @notice Processes and validates the optional Opensea signature, if provided.
+    /// @param order The order parameters in Seaport's format.
+    /// @param openseaSignature The Opensea signature (optional).
+    /// @return orderList An array containing the original order and the Opensea order (if provided).
+    function processOpenseaSignature(Order memory order, bytes memory openseaSignature) internal view returns (Order[] memory orderList) {
+        if (openseaSignature.length > 0) {
+            Order memory openseaOrder = composeOpenseaOrder(order, openseaSignature);
+            orderList = new Order[](2);
+            orderList[0] = order;
+            orderList[1] = openseaOrder;
+        } else {
+            orderList = new Order[](1);
+            orderList[0] = order;
+        }
+        return orderList;
+    }
+
+    /// @notice Stores the order and any additional recipients in the contract and updates the relevant mappings.
+    /// @param order The order parameters in Seaport's format.
+    function storeOrderAndRecipients(Order memory order) internal {
+        address token = order.parameters.offer[0].token;
+        uint256 tokenId = order.parameters.offer[0].identifierOrCriteria;
+
+        (bytes1 v, bytes32 r, bytes32 s) = extractSignatureComponents(order.signature);
+
+        uint8 considerations = uint8(order.parameters.consideration.length);
+        storeAdditionalRecipients(order, considerations);
+
+        UniqueOrderParameters memory uniqueOrderParameters = createUniqueOrderParameters(order, v, r, s);
+        orders[token][tokenId] = uniqueOrderParameters;
+
+        addTokenId(token, tokenId);
+    }
+
+    /// @notice Extracts the v, r, and s components from the signature.
+    /// @param signature The order signature as bytes.
+    /// @return v The 'v' component of the signature.
+    /// @return r The 'r' component of the signature.
+    /// @return s The 's' component of the signature.
+    function extractSignatureComponents( 
+        bytes memory signature 
+    ) internal pure returns (
+        bytes1 v, bytes32 r, bytes32 s
+    ) {
+        assembly {
+            v := mload(add(signature, 0x20))
+            r := mload(add(signature, 0x21))
+            s := mload(add(signature, 0x41))
+        }
+    }
+
+    /// @notice Creates a UniqueOrderParameters struct from the given order and its signature components.
+    /// @param order The order data.
+    /// @param v The 'v' component of the order signature.
+    /// @param r The 'r' component of the order signature.
+    /// @param s The 's' component of the order signature.
+    /// @return A UniqueOrderParameters struct containing a condensed representation of the order data and its signature components.
+    function createUniqueOrderParameters(
+        Order memory order,
+        bytes1 v,
+        bytes32 r,
+        bytes32 s
+    ) internal pure returns (
+        UniqueOrderParameters memory
+    ) {
+        return UniqueOrderParameters({
+            amount: uint64(order.parameters.consideration[0].startAmount / 1e3),
+            startTime: uint32(order.parameters.startTime),
+            endTime: uint32(order.parameters.endTime),
+            salt: uint16(order.parameters.salt),
+            vSignature: v,
+            rSignature: r,
+            sSignature: s
+        });
+    }
+
+    /// @notice Stores additional recipients for an order, if any, in the additionalRecipients mapping.
+    /// @param order The order data containing the additional recipient information.
+    /// @param considerations The number of considerations (recipients) in the order.
+    function storeAdditionalRecipients(
+        Order memory order,
+        uint8 considerations
+    ) internal {
+        if (considerations > 1) {
+            address token = order.parameters.offer[0].token;
+            uint256 tokenId = order.parameters.offer[0].identifierOrCriteria;
+            for (uint8 i = 1; i < considerations; i++) {
+                AdditionalRecipient memory recipient = AdditionalRecipient({
+                    recipient: order.parameters.consideration[i].recipient,
+                    amount: uint64(order.parameters.consideration[i].startAmount / 1e3)
+                });
+                additionalRecipients[token][tokenId][i] = recipient;
+            }
+        }
+    }
+
+    /// @notice Composes an Opensea order based on the provided order and signature.
+    /// @param order The base order, not including Opensea as a recipient.
+    /// @param openseaSignature The signature of the Opensea order.
+    /// @return openseaOrder The composed Opensea order.
+    function composeOpenseaOrder(
+        Order memory order,
+        bytes memory openseaSignature
+    ) internal view returns (
+        Order memory openseaOrder
+    ) {
+        openseaOrder.parameters.offerer = order.parameters.offerer;
+        openseaOrder.parameters.zone = address(0);
+        openseaOrder.parameters.offer = order.parameters.offer;
+        openseaOrder.parameters.orderType = order.parameters.orderType;
+        openseaOrder.parameters.startTime = order.parameters.startTime;
+        openseaOrder.parameters.endTime = order.parameters.endTime;
+        openseaOrder.parameters.zoneHash = order.parameters.zoneHash;
+        openseaOrder.parameters.salt = order.parameters.salt;
+        openseaOrder.parameters.conduitKey = openseaConduitKey;
+        openseaOrder.signature = openseaSignature;
+
+        uint256 considerations = order.parameters.consideration.length;
+
+        uint256 amount;
+        if (considerations == 1) {
+            amount = order.parameters.consideration[0].startAmount*openseaFee/1e18;
+        } else {
+            amount = (order.parameters.consideration[0].startAmount+order.parameters.consideration[1].startAmount)*openseaFee/1e18;
+        }
+
+        ConsiderationItem[] memory openseaConsideration = new ConsiderationItem[](considerations + 1);
+        openseaConsideration[0] = order.parameters.consideration[0];
+        openseaConsideration[1] = ConsiderationItem({
+            itemType: ItemType.NATIVE,
+            token: address(0),
+            identifierOrCriteria: 0,
+            startAmount: amount,
+            endAmount: amount,
+            recipient: payable(openseaAddress)
+        });
+        if (considerations == 2) {
+            openseaConsideration[2] = order.parameters.consideration[1];
+        }
+        openseaOrder.parameters.consideration = openseaConsideration;
+        openseaOrder.parameters.totalOriginalConsiderationItems = order.parameters.totalOriginalConsiderationItems + 1;
+
+        return openseaOrder;
     }
 
     /// @notice Composes a Seaport AdvancedOrder object for a given order.
@@ -170,7 +333,7 @@ contract MarketMap {
             uniqueOrderParameters.rSignature,
             uniqueOrderParameters.sSignature
         );
-        uint256 amount = uint256(uniqueOrderParameters.amount)*1e13;
+        uint256 amount = uint256(uniqueOrderParameters.amount)*1e3;
         OfferItem[] memory offer = new OfferItem[](1);
         offer[0] = OfferItem({
             itemType: ItemType.ERC721,
@@ -180,7 +343,14 @@ contract MarketMap {
             endAmount: 1
         });
 
-        ConsiderationItem[] memory consideration = new ConsiderationItem[](1);
+        uint256 considerations;
+        uint8 i = 1;	
+        while (additionalRecipients[token][tokenId][i].amount > 0) {
+            considerations++;
+            i++;
+        }
+
+        ConsiderationItem[] memory consideration = new ConsiderationItem[](i);
         consideration[0] = ConsiderationItem({
             itemType: ItemType.NATIVE,
             token: address(0),
@@ -189,6 +359,21 @@ contract MarketMap {
             endAmount: amount,
             recipient: payable(offerer)
         });
+
+        if (i > 1) {
+            for (uint8 j = 1; j < i; j++) {
+                AdditionalRecipient memory recipient = additionalRecipients[token][tokenId][j];
+                uint256 recipientAmount = uint256(recipient.amount)*1e3;
+                consideration[j] = ConsiderationItem({
+                    itemType: ItemType.NATIVE,
+                    token: address(0),
+                    identifierOrCriteria: 0,
+                    startAmount: recipientAmount,
+                    endAmount: recipientAmount,
+                    recipient: payable(recipient.recipient)
+                });
+            }
+        }
 
         AdvancedOrder memory advancedOrder = AdvancedOrder({
             parameters: OrderParameters({
@@ -202,7 +387,7 @@ contract MarketMap {
                 zoneHash: bytes32(0),
                 salt: uint256(uniqueOrderParameters.salt),
                 conduitKey: bytes32(0),
-                totalOriginalConsiderationItems: 1
+                totalOriginalConsiderationItems: uint256(i)
             }),
             numerator: 1,
             denominator: 1,
@@ -235,25 +420,5 @@ contract MarketMap {
         tokenIdsWithOrders[token].pop();
 
         delete tokenIdToIndex[token][tokenId];
-    }
-
-    function getOrderStatus(address token, uint256 tokenId) public view returns (bool isActive) {
-        AdvancedOrder memory advancedOrder = composeOrder(token, tokenId);
-        OrderComponents memory orderComponents = OrderComponents({
-            offerer: advancedOrder.parameters.offerer,
-            zone: advancedOrder.parameters.zone,
-            offer: advancedOrder.parameters.offer,
-            consideration: advancedOrder.parameters.consideration,
-            orderType: advancedOrder.parameters.orderType,
-            startTime: advancedOrder.parameters.startTime,
-            endTime: advancedOrder.parameters.endTime,
-            zoneHash: advancedOrder.parameters.zoneHash,
-            salt: advancedOrder.parameters.salt,
-            conduitKey: advancedOrder.parameters.conduitKey,
-            counter: seaport.getCounter(advancedOrder.parameters.offerer)
-        });
-        bytes32 orderHash = seaport.getOrderHash(orderComponents);
-        (isActive,,,) = seaport.getOrderStatus(orderHash);
-        return isActive;
     }
 }